// src/types/telegram.d.ts
  type ParseMode = 'Markdown' | 'MarkdownV2' | 'HTML';
// ...
  interface SendBasicOptions {
    message_thread_id?: number | undefined;
    disable_notification?: boolean | undefined;
    reply_to_message_id?: number | undefined;
    reply_markup?:
      | InlineKeyboardMarkup
      | ReplyKeyboardMarkup
      | ReplyKeyboardRemove
      | ForceReply
      | undefined;
    protect_content?: boolean | undefined;
    allow_sending_without_reply?: boolean | undefined;
  }

  interface SendMessageOptions extends SendBasicOptions {
    parse_mode?: ParseMode | undefined;
    disable_web_page_preview?: boolean | undefined;
  }
// ...
  interface ReplyKeyboardMarkup {
    keyboard: KeyboardButton[][];
    is_persistent?: boolean | undefined;
    resize_keyboard?: boolean | undefined;
    one_time_keyboard?: boolean | undefined;
    input_field_placeholder?: string | undefined;
    selective?: boolean | undefined;
  }
// ...
  interface ReplyKeyboardRemove {
    remove_keyboard: boolean;
    selective?: boolean | undefined;
  }

  interface InlineKeyboardMarkup {
    inline_keyboard: InlineKeyboardButton[][];
  }
// ...
  interface ForceReply {
    force_reply: boolean;
    input_field_placeholder?: string | undefined;
    selective?: boolean | undefined;
  }